+++ domain/src/commonMain/kotlin/co/lucaspinazzola/example/domain/interactor/gif/UpdateGifsUseCase.kt
@@ -5,7 +5,7 @@ import co.lucaspinazzola.example.domain.repo.GiphyRepository
 
 class UpdateGifsUseCase(private val giphyRepository: GiphyRepository) {
 
-    suspend fun execute(query: String, offset: Int)  {
+    suspend fun execute(query: String, offset: Long)  {
         giphyRepository.updateGifs(query, offset)
     }
