+++ domain/src/commonMain/kotlin/co/lucaspinazzola/example/domain/repo/GiphyRepository.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 interface GiphyRepository {
 
     suspend fun getGifs(): List<Gif>
-    suspend fun updateGifs(query: String, offset: Int)
+    suspend fun updateGifs(query: String, offset: Long)
 
     fun listenForGifUpdates(onChangePublisherSubscribed: suspend ()->Unit) : Flow<List<Gif>>
